Fix bug preventing disabling
Funny business with storing and remembering the path to the CSS file was keeping the extension from properly disabling itself.
Also, the flag indicating that the CSS had been enabled was not being set on first run, requiring you to press the button twice to disable it the first time (if you were able to disable it at all before).
Both issues fixed, and you can now properly toggle the CSS and effects on and off however many times you want!